requirements.txt
asyncpg==0.24.0
asyncpg-migrate==0.0.7
psycopg2-binary==2.9.1
python-dotenv==0.19.0
uvloop==0.14.0